feat(review): refactor package structure by consolidating aggregation helpers into a single package
