+++ processing/src/main/java/pt/estga/processing/services/similarity/EmbeddingPreprocessor.java
@@ -1,4 +1,4 @@
-package pt.estga.processing.services.similarity.helpers;
+package pt.estga.processing.services.similarity;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
