+++ processing/src/main/java/pt/estga/processing/services/similarity/ParityChecker.java
@@ -1,4 +1,4 @@
-package pt.estga.processing.services.similarity.helpers;
+package pt.estga.processing.services.similarity;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
