+++ processing/src/main/java/pt/estga/processing/services/similarity/aggregation/ScoreCalculator.java
@@ -1,4 +1,4 @@
-package pt.estga.processing.services.similarity.helpers.aggregation;
+package pt.estga.processing.services.similarity.aggregation;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
